requirements.txt
# Python 3.14 совместимые зависимости
# Без жёстких пинов — pip сам подберёт совместимые версии

# Bot — без пина на aiogram чтобы не конфликтовал с pydantic>=2.10
aiogram>=3.13.1
aiofiles>=24.1.0
aiohttp>=3.9.0,<3.11

# Web / Admin
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
jinja2>=3.1.0
itsdangerous>=2.2.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
greenlet>=3.1.1

# Cache / State
redis>=5.0.0

# Media
pillow>=10.0.0

# Payments
cryptography>=43.0.0

# Config — pydantic>=2.10 нужен для Python 3.14 (pyo3>=0.23)
pydantic>=2.10.0
pydantic-settings>=2.8.0
python-dotenv>=1.0.0
APScheduler>=3.10.0

# Monitoring
prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=7.0.0

# Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
aiosqlite>=0.19.0
